+++ p2panda-rs/src/document/mod.rs
@@ -252,11 +252,14 @@
 //! # Ok(())
 //! # }
 //! ```
+
 #[allow(clippy::module_inception)]
 mod document;
 mod document_view;
 mod error;
 
+#[allow(unused_imports)]
+use document::{build_graph, reduce};
 pub use document::{Document, DocumentBuilder};
-pub use document_view::DocumentView;
+pub use document_view::{DocumentView, DocumentViewId};
 pub use error::{DocumentBuilderError, DocumentError, DocumentViewError};